+++ src/main/java/org/opensearch/knn/index/query/iterators/ByteVectorIdsKNNIterator.java
@@ -6,8 +6,6 @@
 package org.opensearch.knn.index.query.iterators;
 
 import org.apache.lucene.search.DocIdSetIterator;
-import org.apache.lucene.util.BitSet;
-import org.apache.lucene.util.BitSetIterator;
 import org.opensearch.common.Nullable;
 import org.opensearch.knn.index.SpaceType;
 import org.opensearch.knn.index.vectorvalues.KNNByteVectorValues;
@@ -21,7 +19,7 @@ import java.io.IOException;
  * The class is used in KNNWeight to score all docs, but, it iterates over filterIdsArray if filter is provided
  */
 public class ByteVectorIdsKNNIterator implements KNNIterator {
-    protected final BitSetIterator bitSetIterator;
+    protected final DocIdSetIterator filterIdsIterator;
     protected final float[] queryVector;
     protected final KNNByteVectorValues byteVectorValues;
     protected final SpaceType spaceType;
@@ -29,12 +27,12 @@ public class ByteVectorIdsKNNIterator implements KNNIterator {
     protected int docId;
 
     public ByteVectorIdsKNNIterator(
-        @Nullable final BitSet filterIdsBitSet,
+        @Nullable final DocIdSetIterator filterIdsIterator,
         final float[] queryVector,
         final KNNByteVectorValues byteVectorValues,
         final SpaceType spaceType
     ) throws IOException {
-        this.bitSetIterator = filterIdsBitSet == null ? null : new BitSetIterator(filterIdsBitSet, filterIdsBitSet.length());
+        this.filterIdsIterator = filterIdsIterator;
         this.queryVector = queryVector;
         this.byteVectorValues = byteVectorValues;
         this.spaceType = spaceType;
@@ -89,10 +87,10 @@ public class ByteVectorIdsKNNIterator implements KNNIterator {
     }
 
     protected int getNextDocId() throws IOException {
-        if (bitSetIterator == null) {
+        if (filterIdsIterator == null) {
             return byteVectorValues.nextDoc();
         }
-        int nextDocID = this.bitSetIterator.nextDoc();
+        int nextDocID = this.filterIdsIterator.nextDoc();
         // For filter case, advance vector values to corresponding doc id from filter bit set
         if (nextDocID != DocIdSetIterator.NO_MORE_DOCS) {
             byteVectorValues.advance(nextDocID);